requirements.txt
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
django-filter==2.4.0
djoser==2.1.0
python-dotenv==0.21.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==4.7.2
requests==2.26.0
sorl-thumbnail==12.9.0



